calc remaining days

+++ app/holiday.js
@@ -4,9 +4,31 @@ import 'jquery-ui/ui/effect';
 import Transition from 'transition';
 import Request from 'request';
 import holidayListTemplate from './templates/holiday/holiday-row.hbs';
+import LoadingScreen from './loading';
 
 $(function () {
 
+    Request.submitRequest(
+        {
+            url: '/api/user/remaining',
+            method: 'GET'
+        }
+    );
+
+    var _validator = {
+        request: {
+            date_from: /^[0-9\/]+/,
+            date_to: function ( val ) {
+                if ( ! val ){ return false; }
+                var from_time = new Date($('form#request').find('input#date_from').first().val()).getTime();
+                var to_time = new Date(val).getTime();
+                return to_time > from_time;
+            }
+        }
+    };
+
+    var loadingScreen = new LoadingScreen();
+    
     var _newId;
     window.Handlebars.registerHelper( 'formatDate', function ( date ) {
         return new Intl.DateTimeFormat().format(new Date(date));
@@ -33,25 +55,28 @@ $(function () {
 
                 // set up actions
                 $( "i#action-delete" ).click( ( e ) => {
-                    var row = $(e.target).parents('tr');
-                    var id = $(e.target).attr('data-holiday-id');
+                    if ( confirm("Are you sure you would like to delete this holiday?") ){
+                        var row = $(e.target).parents('tr');
+                        var id = $(e.target).attr('data-holiday-id');
 
-                    Request.submitRequest(
-                        {
-                            url: `/api/holiday/${id}`,
-                            method: 'DELETE'
-                        }
-                    ).then(
-                        function () {
-                            row.remove();
-                        }
-                    );
+                        Request.submitRequest(
+                            {
+                                url: `/api/holiday/${id}`,
+                                method: 'DELETE'
+                            }
+                        ).then(
+                            function () {
+                                row.remove();
+                            }
+                        );
+                    }
                 });
                 $( "i#action-edit" ).click( ( e ) => {
                     console.log( "EDIT", $(e.target).attr('data-holiday-id') );
                 });
             }
         );
+
     };
 
     var transition = new Transition();
@@ -62,27 +87,41 @@ $(function () {
     });
 
     $( "button#request" ).click( () => {
-        transition.navigate( "sending" );
+        loadingScreen.show();
         Request.submitRequest(
             {
                 url: '/api/holiday',
-                form: 'request'
+                form: 'request',
+                validate: _validator.request
             }
         ).then(
             function ( data ){
+                $('form#request > div#error').html('');
                 _newId = data.id;
                 transition.navigate("current");
+                loadingScreen.hide();
             },
-            function ( err ) {
-                $('#view-container-2 > div').html(
-                    "<h3>"+err+"</h3>" +
-                    "<p>Something bad has happened," +
-                    "<a href='#' onclick='makeTransition(\"view\")'>go back</a>"
+            function ( error ) {
+                $('form#request > div#error').html(
+                    "<strong>Error</strong>" + "<p>"+error+"</p>"
                 );
+                loadingScreen.hide();
             }
         );
     });
 
-    $( "input#date_from, input#date_to" ).datepicker();
+    /* thanks http://stackoverflow.com/a/3827570 */
+    var customRange = function ( input ) {
+        if (input.id == 'date_to') {
+            return {
+                minDate: new Date($('#date_from').val())
+            };
+        }
+    };
+
+    $( "input#date_from, input#date_to" ).datepicker({
+        constrainInput: true,
+        beforeShow: customRange
+    });
 
 });
